+++ classes/local/session/data_access/analytics_vault_interface.php
@@ -30,15 +30,6 @@ defined('MOODLE_INTERNAL') || die();
 interface analytics_vault_interface {
 
     /**
-     * Get Session Recency in days for a particular student
-     *
-     * @param int   $courseid   The course id in reference
-     * @param int   $studentid  The student id in reference
-     * @return int  $days       The number of days since last session
-     */
-    public static function get_session_recency(int $courseid, int $studentid);
-
-    /**
      * Get course activity for a student from the logs.
      *
      * @param int   $courseid   The course id in reference