+++ classes/local/session/data_access/booking_vault.php
@@ -276,12 +276,9 @@ class booking_vault implements booking_vault_interface {
     public static function get_last_booked_session(int $courseid, int $userid, bool $isinstructor = false) {
         global $DB;
 
-        $sql = 'SELECT s.starttime as lastbookedsession
+        $sql = 'SELECT MAX(b.timemodified) as lastbookedsession
                 FROM {' . static::DB_BOOKINGS. '} b
-                INNER JOIN {' . static::DB_SLOTS. '} s ON s.id = b.slotid
-                WHERE b.courseid = :courseid AND ' . ($isinstructor ? 'b.userid = :userid' : 'b.studentid = :userid') . '
-                ORDER BY b.timemodified DESC
-                LIMIT 1';
+                WHERE b.courseid = :courseid AND ' . ($isinstructor ? 'b.userid = :userid' : 'b.studentid = :userid');
 
         return $DB->get_record_sql($sql, ['courseid'=>$courseid, 'userid'=>$userid]);
     }
@@ -300,7 +297,8 @@ class booking_vault implements booking_vault_interface {
                 FROM {' . static::DB_BOOKINGS. '}
                 WHERE courseid = :courseid AND ' . ($isinstructor ? 'userid = :userid' : 'studentid = :userid');
 
-        return $DB->get_record_sql($sql, ['courseid'=>$courseid, 'userid'=>$userid])->totalsessions;
+        $sessions = $DB->get_record_sql($sql, ['courseid'=>$courseid, 'userid'=>$userid]);
+        return !empty($sessions) ? $sessions->totalsessions : 0;
     }
 
     /**